requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.2.1
Django==2.1.7
django-crispy-forms==1.7.0
django-markdown-deux==1.0.5
django-pagedown==1.0.4
enum34==1.1.6
gunicorn==19.9.0
ipython==5.5.0
ipython-genutils==0.2.0
markdown2==2.3.5
pathlib2==2.3.0
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2018.9
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
dj-database-url==0.5.0
psycopg2==2.7.6.1
whitenoise==4.1.2
Pillow==6.0.0
